third_party/WebKit/LayoutTests/http/tests/cache/smil-usecounter-in-cached-image.html

<!DOCTYPE html>
<script src="../../js-test-resources/js-test.js"></script>
<script>
var SVGSMILAnimationInImageRegardlessOfCache = 768; // From UseCounter.h
window.jsTestIsAsync = true;

window.onload = function() {
    var img = new Image();
    if (location.hash != '#loadWithCachedImage') {
        img.onload = function() {
            location.hash = 'loadWithCachedImage';
            window.location.reload();
        }
    } else {
        shouldBeFalse("window.internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache)");
        img.onload = function() {
            shouldBeTrue("window.internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache)");
            finishJSTest();
        }
    }
    img.src = "resources/svg-with-animation.svg";
}
</script>
</html>
