third_party/WebKit/LayoutTests/http/tests/cache/smil-usecounter-in-cached-image-expected.txt

PASS window.internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache) is false
PASS window.internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache) is true
PASS successfullyParsed is true

TEST COMPLETE
